Add `Proxy typ` param to `dataShape` signature
The redundant `Proxy typ` parameter essentially introduces `typ` into
the signature body, making the compiler less confused and able to
solve the contraints within a finite number of cycles.

GHC 9.4.x caused the original signature to trigger an infinite loop in
the constraint solver. This led to the offending module containing this
signature to compile indefinitely.