no-op: Add lower bounds CI test.
One issue is that ghcup doesn't support GHC versions earlier than 8.4,
while the current lower bounds are so early that some of them won't
build on 8.4 (presumably they will build on something earlier, but I
haven't tested that far back).

Therefore, this doesn't _really_ test the absolute lower bounds. It
just goes as far back as is easy to test in CI using the standard
haskell-actions github actions.

(I'm calling this a "no-op" because it doesn't affect the code at
all.)